initial integration of validate test into cli
